test(primitives): move tests to a dedicated file per review suggestion
Pure move of the inline #[cfg(test)] module to src/test.rs, matching the
errors/storage/auth module layout. No test content changes.
